+++ Sources/Navigator/Audiobook/AudioNavigator.swift
@@ -27,11 +27,12 @@ open class _AudioNavigator: _MediaNavigator, _AudioSessionUser, Loggable {
     public init(publication: Publication, initialLocation: Locator? = nil) {
         self.publication = publication
         self.initialLocation = initialLocation
-            ?? publication.readingOrder.first.map { Locator(link: $0) }
+            ?? publication.readingOrder.first.flatMap { publication.locate($0) }
         
         let durations = publication.readingOrder.map { $0.duration ?? 0 }
+        let totalDuration = durations.reduce(0, +)
+        
         self.durations = durations
-        let totalDuration = publication.metadata.duration ?? durations.reduce(0, +)
         self.totalDuration = (totalDuration > 0) ? totalDuration : nil
     }
     
@@ -255,7 +256,10 @@ open class _AudioNavigator: _MediaNavigator, _AudioSessionUser, Loggable {
 
     @discardableResult
     public func go(to link: Link, animated: Bool = false, completion: @escaping () -> Void = {}) -> Bool {
-        return go(to: Locator(link: link), animated: animated, completion: completion)
+        guard let locator = publication.locate(link) else {
+            return false
+        }
+        return go(to: locator, animated: animated, completion: completion)
     }
     
     @discardableResult